+++ src/common-select/common-select.jsx
@@ -53,6 +53,8 @@ type CommonSelectProps = {
   isClearable?: boolean,
   /** Disables the select from being used */
   isDisabled?: boolean,
+  /** Renders a LinearProgress bar beneath the common select input */
+  isLoading?: boolean,
   /** The text displayed in the input before the user begins typing */
   placeholder?: string,
   /**